Feature: Allow RTOS CPU yielding
The biggest advantage of using a RTOS (FreeRTOS, Zephyr, …) is their
ability to hide the complexity of asynchronous systems, but this often
requires cooperation by the threads.

This library makes extensive use of busy waiting when it is being used
in blocking mode, which in the way it was implemented broke the
cooperation principle. To solve this this patch introduces a function
named `Mcb_RelinquishCPU` to `mcb_usr.h`.

This function is executed on every iteration in the libraries busy
waiting loops. When using an RTOS, the user can override it in order to
execute their native thread yield / wait / … functions and allow the
OS reschedule the task.

On non RTOS systems this is of course a no-op.

Signed-off-by: Elisabeth Friedrich <auxsys@flowerpot.me>